CARD-44: review fixes

+++ src/kafka/payment/schemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel
@@ -5,4 +6,4 @@ from pydantic import BaseModel
 
 class UpdateBalance(BaseModel):
     user_id: UUID
-    balance: int
+    balance: Decimal
